+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of lizard-box
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added javascript-loaded boxes, for when you need popups that are
+  big.
 
 
 0.1 (2012-08-01)